scanner_QR_code/requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
pyzbar==0.1.8
requests==2.22.0
urllib3==1.25.3
